storage/Storage: refactor GetRecordBatch to be blocking
This commit makes GetRecordBatch blocking; if the given offset does not
yet exist, GetRecordBatch will block until either it exists or until the
given context expires. This allows callers to make blocking calls
instead of retrying until it becomes available. This should be both
simpler for the caller and reduce the load on the sever.