sebtopic/s3: fix s3 errors not being propagated
This commit fixes an error where the return value of calls to Close()
were not used. This was a major problem, since files were uploaded to S3
in a call to S3Storage writer's Close().

An easy way to recreate the bug was to start Seb with credentials that
were valid only for long enough to initialize a given topic (allowing
Seb to list files in order to understand the current state) and then add
a new record once the credentials had expired.