all/read path: require pre-allocated Batch as input
NOTE: this changes the interface of the GET /records endpoint

- sebrecords/Batch: add Append()
- sebtopic/ReadRecords: use Batch.Append() to simplify logic
- records/Parse: take pre-allocated Batch as input
- all: move errors seb -> seberrors
- sebrecords: refactor Batch.IndividualRecords()
- http,client: use 'raw data with offsets' instead of 'part per record' format
- cmd/app: configure batch pools

BUG: client/GetRecords: invalid reuse of buffers
The bug was caused by buffers used in MultipartFormDataToRecords()
being returned to the shared pool within MultipartFormDataToRecords,
before their usage truly ended. The acquired buffer's underlying byte
slice was used to instantiate a sebrecords.Batch, which was passed to
the caller of MultipartFormDataToRecords(). At this point the caller of
MultipartFormDataToRecords() had a reference to the byte slice via the
returned sebrecords.Batch while the buffer was back in the pool and
another caller could retrieve it, overriding the contents of the
underlying byte slice within the sebrecords.Batch.